Updated 20230422

+++ hw4/Attendance.cpp
@@ -1,5 +1,5 @@
 /*  鄒雨笙 F64081070 程式設計二 第4次作業 2023/4/26
-    
+
 */
 #include <iostream>
 #include "Attendance.h"
@@ -12,12 +12,9 @@ void Attendance::setRecord(int conWorkday, std::string &startDay, std::string &e
 }
 
 void Attendance::printRecord() {
-    for (int i = 0; i < conDayRecord.size(); i++) {
+    for (size_t i = 0; i < conDayRecord.size(); i++) {
         cout << conDayRecord[i].getConWorkday() << "," << conDayRecord[i].getStartDay() << "," << conDayRecord[i].getEndDay() << endl;
     }
 
 }
 
-        
-        
-
